requirements.txt
psutil>=5.9.0
flask>=2.3.0
flask-socketio>=5.3.0
pystray>=0.19.0
pillow>=10.0.0
pywin32>=306
aiohttp>=3.8.0
python-socketio>=5.8.0
requests>=2.31.0
dnspython>=2.4.0
schedule>=1.2.0
